Turn aissrc into a live source.
This fixes the inability to mux raw images into mp4s (or any
other container format) for streams whose elements are user populated
RawImages (note that this was never a problem with raw images from a
video/camera).

The issue is that a non-live src element does not assign pts' to
incoming buffers, but muxers depend on it during container assembly.
It was fine to omit pts for raw images from video file sources,
since muxers may use the framerate to generate the pts internally.

Live sources on the other hand receives a clock synchronized to the
pipeline. It is then able to attach pts' to buffers based on that clock.

Change-Id: I4b293f4f71a72792ad6df134de8bf590ea87aa6b