A class whose instances write video files.
GstreamerVideoWriter writes a video file from the sequence of
GstreamerBuffers that it is given. We may write video files from
Packets directly in the future (which will include all the packet
type validations), but we can start with this interface first.

Each instances writes a single file; chunking of video boundaries from
the stream is up to the client.

Change-Id: Ib5ca8891681259a3cb9cdbf2620559730f7a1e79