An exporter of video files.
The exporter receives packets from a stream and produces a sequence of
video files.

This CL just saves the files to the present directory. GCS upload
support will come in a later CL.

This CL also just accepts a stream server's packets. Using a local
gstreamer pipeline as the input source will come in a later CL.

Change-Id: I6a10d8913835e3bb36f5e27f4fd387c214cd7589